+++ Common/Distance.h
@@ -7,7 +7,7 @@
 using namespace std;
 
 //Function that calculates the manhattan distance between two images
-double manhattanDistance(vector<unsigned char> * firstImagePixels, vector<unsigned char> * secondImagePixels);
+int manhattanDistance(vector<unsigned char> * firstImagePixels, vector<unsigned char> * secondImagePixels);
 
 
 #endif //PROJECT_TASK1_DISTANCE_H
